+++ Dx3D/Dx3D/003.Object/cSkinnedMesh.cpp
@@ -467,7 +467,7 @@ float cSkinnedMesh::GetCurPos()
         D3DXTRACK_DESC desc;
         m_pAnimController->GetTrackDesc(0, &desc);
 
-        float CurPos = desc.Position / pAnimSet->GetPeriod();
+        float CurPos = (float)desc.Position / (float)pAnimSet->GetPeriod();
 
         nResult =  CurPos;
     }
